Suppress sign-conversion warnings in MinGW environment

+++ src/quant.c
@@ -747,8 +747,8 @@ computeHistogram(unsigned char const    /* in */  *data,
         }
     }
 
-    colorfreqtableP->size = ref - refmap;
-    status = alloctupletable(&colorfreqtableP->table, depth, ref - refmap, allocator);
+    colorfreqtableP->size = (unsigned int)(ref - refmap);
+    status = alloctupletable(&colorfreqtableP->table, depth, (unsigned int)(ref - refmap), allocator);
     if (SIXEL_FAILED(status)) {
         goto end;
     }
